Day 27
I spent most of the day revisiting clean architecture principles.

I created a cubit and state for the profile details. It's interesting how many lines of code just to create this.

Creating a repository abstract class => implementing that repository => creating a firebase user sources abstract class => implementing the firebase user sources class => creating a userEntity => creating a user model including a fromJson function and a to userEntity extention function => creating a get user usecase extending the base usecase class => registering that usecase in the service locator => then creating the UI with bloc builder and bloc provider.

And this will be repeated with every single feature 🙂